refactor: 동일한 폴더에 테스트 코드를 옮겨 응집도 향상
